Break code into smaller files
This is ahead of some work to be able to handle
different types of instruction sets and trees.
