+++ src/lib.rs
@@ -3,377 +3,15 @@
 extern crate rlp;
 extern crate sha3;
 
+pub mod instruction;
+pub mod multiproof;
+pub mod node;
 pub mod utils;
 
-use sha3::{Digest, Keccak256};
+use instruction::*;
+use multiproof::*;
+use node::*;
 use utils::*;
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum Node {
-    Hash(Vec<u8>), // (Hash, # empty spaces)
-    Leaf(NibbleKey, Vec<u8>),
-    Extension(NibbleKey, Box<Node>),
-    Branch(Vec<Node>),
-    EmptySlot,
-}
-
-impl Default for Node {
-    fn default() -> Self {
-        Node::EmptySlot
-    }
-}
-
-impl rlp::Encodable for Node {
-    fn rlp_append(&self, s: &mut rlp::RlpStream) {
-        match self {
-            Node::Leaf(ref k, ref v) => {
-                s.append_list::<Vec<u8>, Vec<u8>>(&vec![k.with_hex_prefix(true), v.to_vec()]);
-            }
-            Node::Extension(ref ext, box node) => {
-                let extension_branch_hash = node.composition();
-
-                s.append_list::<Vec<u8>, Vec<u8>>(&vec![
-                    ext.with_hex_prefix(false),
-                    extension_branch_hash,
-                ]);
-            }
-            Node::Branch(ref nodes) => {
-                let mut stream = rlp::RlpStream::new();
-                stream.begin_unbounded_list();
-                for node in nodes {
-                    let hash = node.composition();
-                    if hash.len() < 32 && hash.len() > 0 {
-                        stream.append_raw(&hash, hash.len());
-                    } else {
-                        stream.append(&hash);
-                    }
-                }
-                // 17th element
-                stream.append(&"");
-                stream.complete_unbounded_list();
-                let encoding = stream.out();
-                s.append_raw(&encoding, encoding.len());
-            }
-            _ => panic!("Not supported yet!"),
-        }
-    }
-}
-
-impl rlp::Decodable for Node {
-    fn decode(rlp: &rlp::Rlp) -> Result<Self, rlp::DecoderError> {
-        if !rlp.is_list() {
-            return Err(rlp::DecoderError::RlpExpectedToBeList);
-        }
-        let keyval = rlp.as_list::<Vec<u8>>()?;
-        let key_bytes = utils::ByteKey(keyval[0].clone());
-        let key_nibbles = NibbleKey::from(key_bytes);
-        // TODO: remove indicator prefix if node is a leaf or extension
-        Ok(Node::Leaf(key_nibbles, keyval[1].clone()))
-    }
-}
-
-// Implement sub-node access based on a nibble key prefix.
-impl std::ops::Index<&NibbleKey> for Node {
-    type Output = Node;
-    #[inline]
-    fn index(&self, k: &NibbleKey) -> &Node {
-        // If the key has 0-length, then the search is over and
-        // the current node is returned.
-        if k.len() == 0 {
-            self
-        } else {
-            match self {
-                Node::Branch(ref children) => {
-                    children[k[0] as usize].index(&NibbleKey::from(&k[1..]))
-                }
-                Node::Extension(ref ext, box child) => {
-                    let factor_length = ext.factor_length(k);
-                    // If the factorized length is that of the extension,
-                    // then the indexing key is compatible with this tree
-                    // and the function recurses.
-                    if factor_length == ext.len() {
-                        child.index(&k[factor_length..].into())
-                    } else {
-                        // Otherwise, we either have a key that is shorter
-                        // or a differing entry. The former returns the
-                        // current node and the latter panics.
-                        if k.len() == factor_length {
-                            self
-                        } else {
-                            panic!("Requested key isn't present in the tree")
-                        }
-                    }
-                }
-                Node::Leaf(ref key, _) => {
-                    let factor_length = key.factor_length(k);
-                    if k[..factor_length] == key[..factor_length] {
-                        self
-                    } else {
-                        panic!("Requested key isn't present in the tree")
-                    }
-                }
-                Node::EmptySlot | Node::Hash(_) => {
-                    panic!("Requested key isn't present in the tree")
-                }
-            }
-        }
-    }
-}
-
-impl Node {
-    pub fn graphviz(&self) -> String {
-        let (nodes, refs) = self.graphviz_rec(NibbleKey::from(vec![]), String::from("root"));
-        format!(
-            "digraph D {{
-\tnode [shape=\"box\",label=\"hash\"];
-
-\t{}
-
-\t{}
-}}",
-            nodes.join("\n\t"),
-            refs.join("\n\t")
-        )
-    }
-
-    fn graphviz_key(key: NibbleKey) -> String {
-        let mut ret = String::new();
-        let nibbles: Vec<u8> = key.into();
-        for (i, nibble) in nibbles.iter().enumerate() {
-            ret.push_str(&format!("<td port=\"{}\">{:x}</td>", i, nibble));
-        }
-        ret
-    }
-
-    fn graphviz_rec(&self, prefix: NibbleKey, root: String) -> (Vec<String>, Vec<String>) {
-        let pref: Vec<u8> = prefix.clone().into();
-        match self {
-            Node::Leaf(ref k, ref v) => {
-                return (
-                    vec![
-                        format!("leaf{} [shape=none,margin=0,label=<<table border=\"0\" cellborder=\"1\" cellspacing=\"0\" cellpadding=\"4\"><tr>{}<td>{}</td></tr></table>>]", hex::encode(pref.clone()), Node::graphviz_key(k.clone()), hex::encode(v)),
-                    ],
-                    vec![format!("{} -> leaf{}", root, hex::encode(pref))],
-                );
-            }
-            Node::Branch(ref subnodes) => {
-                let name = format!("branch{}", hex::encode(pref.clone()));
-                let mut label = String::from("");
-                for i in 0..15 {
-                    label.push_str(&format!("<td port=\"{}\">{:x}</td>", i, i));
-                }
-                println!("{:?}", hex::encode(pref.clone()));
-                let mut refs = if prefix.len() > 0 {
-                    vec![format!("{} -> {}", root, name)]
-                } else {
-                    Vec::new()
-                };
-                let mut nodes = vec![format!("{} [shape=none,label=<<table border=\"0\" cellborder=\"1\" cellspacing=\"0\"><tr>{}</tr></table>>]", name, label)];
-                for (i, subnode) in subnodes.iter().enumerate() {
-                    let mut subkey: Vec<u8> = prefix.clone().into();
-                    subkey.push(i as u8);
-                    let (sn, sr) = Node::graphviz_rec(
-                        &subnode,
-                        NibbleKey::from(subkey),
-                        format!("{}:{}", name, i),
-                    );
-                    nodes.extend(sn);
-                    refs.extend(sr);
-                }
-                return (nodes, refs);
-            }
-            Node::Extension(ref ext, ref subnode) => {
-                let name = format!("extension{}", hex::encode(pref.clone()));
-                let mut subkey: Vec<u8> = prefix.clone().into();
-                subkey.extend_from_slice(&ext[0..]);
-                let (mut sn, mut sr) = Node::graphviz_rec(
-                    subnode,
-                    NibbleKey::from(subkey),
-                    format!("{}:{}", name, ext.len() - 1),
-                );
-                sn.push(format!("{} [shape=none,label=<<table border=\"0\" cellspacing=\"0\" cellborder=\"1\"><tr>{}</tr></table>>]", name, Node::graphviz_key(ext.clone())));
-                sr.push(format!("{} -> {}:{}", root, name, 0));
-                return (sn, sr);
-            }
-            Node::Hash(ref h) => {
-                let name = format!("hash{}", hex::encode(pref.clone()));
-                let label = if prefix.len() > 0 {
-                    String::from("hash")
-                } else {
-                    String::from("root")
-                };
-                return (
-                    vec![format!("{} [label=\"{}\"]", name, label)],
-                    vec![format!("{} -> {}", root, name)],
-                );
-            }
-            _ => {
-                /* Ignore EmptySlot */
-                (vec![], vec![])
-            }
-        }
-    }
-
-    pub fn is_key_present(&self, key: &NibbleKey) -> bool {
-        match self {
-            Node::Leaf(ref k, _) => k == key,
-            Node::Hash(_) => false,
-            Node::Branch(ref children) => {
-                children[key[0] as usize].is_key_present(&NibbleKey::from(&key[1..]))
-            }
-            Node::Extension(ref ext, box child) => {
-                ext.len() <= key.len()
-                    && *ext == NibbleKey::from(&key[..ext.len()])
-                    && child.is_key_present(&NibbleKey::from(&key[ext.len()..]))
-            }
-            Node::EmptySlot => false,
-        }
-    }
-
-    pub fn hash(&self) -> Vec<u8> {
-        let composed_node = self.composition();
-
-        // If `composition` returned a payload whose length is less
-        // than 32 bytes, then compute its keccack256 hash in order
-        // to return the root hash.
-        if composed_node.len() < 32 {
-            let mut hasher = Keccak256::new();
-            hasher.input(&composed_node);
-            Vec::<u8>::from(&hasher.result()[..])
-        } else {
-            composed_node
-        }
-    }
-
-    // This is the composition function that is described as `n` in the
-    // yellow paper. The difference with the root of the merkle tree is
-    // that it can potentially return an array of bytes whose length is
-    // less than 32 bytes, in which case is represents the RLP encoding
-    // of the root node.
-    fn composition(&self) -> Vec<u8> {
-        use Node::*;
-        match self {
-            EmptySlot => Vec::new(),
-            Hash(h) => h.to_vec(),
-            _ => {
-                let encoding = rlp::encode(self);
-
-                // Only hash if the encoder output is more than 32 bytes.
-                if encoding.len() >= 32 {
-                    let mut hasher = Keccak256::new();
-                    hasher.input(&encoding);
-                    Vec::<u8>::from(&hasher.result()[..])
-                } else {
-                    encoding
-                }
-            }
-        }
-    }
-}
-
-const BRANCH_OPCODE: usize = 0;
-const HASHER_OPCODE: usize = 1;
-const LEAF_OPCODE: usize = 2;
-const EXTENSION_OPCODE: usize = 3;
-const ADD_OPCODE: usize = 4;
-
-#[derive(Debug, PartialEq)]
-pub enum Instruction {
-    BRANCH(usize),
-    HASHER,
-    LEAF(usize),
-    EXTENSION(Vec<u8>),
-    ADD(usize),
-}
-
-impl rlp::Encodable for Instruction {
-    fn rlp_append(&self, s: &mut rlp::RlpStream) {
-        match self {
-            Instruction::EXTENSION(ref ext) => {
-                s.begin_list(2).append(&EXTENSION_OPCODE).append_list(ext)
-            }
-            Instruction::BRANCH(size) => s.begin_list(2).append(&BRANCH_OPCODE).append(size),
-            Instruction::HASHER => s.begin_list(1).append(&HASHER_OPCODE),
-            Instruction::LEAF(size) => s.begin_list(2).append(&LEAF_OPCODE).append(size),
-            Instruction::ADD(index) => s.begin_list(2).append(&ADD_OPCODE).append(index),
-        };
-    }
-}
-
-impl rlp::Decodable for Instruction {
-    fn decode(rlp: &rlp::Rlp) -> Result<Self, rlp::DecoderError> {
-        let instrrlp = rlp.at(0usize)?;
-        let instr: usize = instrrlp.as_val()?;
-
-        if instr >= 5 {
-            return Err(rlp::DecoderError::Custom("Invalid instruction opcode {}"));
-        }
-
-        if instr == EXTENSION_OPCODE {
-            Ok(Instruction::EXTENSION(rlp.at(1)?.as_list()?))
-        } else if instr == HASHER_OPCODE {
-            Ok(Instruction::HASHER)
-        } else {
-            let size: usize = rlp.at(1usize)?.as_val()?;
-            let i = match instr {
-                BRANCH_OPCODE => Instruction::BRANCH(size),
-                HASHER_OPCODE => Instruction::HASHER,
-                LEAF_OPCODE => Instruction::LEAF(size),
-                ADD_OPCODE => Instruction::ADD(size),
-                _ => panic!("This should never happen!"), /* Famous last words */
-            };
-
-            Ok(i)
-        }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub struct Multiproof {
-    pub hashes: Vec<Vec<u8>>,           // List of hashes in the proof
-    pub instructions: Vec<Instruction>, // List of instructions in the proof
-    pub keyvals: Vec<Vec<u8>>,          // List of RLP-encoded (key, value) pairs in the proof
-}
-
-impl rlp::Encodable for Multiproof {
-    fn rlp_append(&self, s: &mut rlp::RlpStream) {
-        s.begin_list(3);
-        s.append_list::<Vec<u8>, Vec<u8>>(&self.hashes[..]);
-        s.append_list::<Vec<u8>, Vec<u8>>(&self.keyvals[..]);
-        s.append_list::<Instruction, Instruction>(&self.instructions[..]);
-    }
-}
-
-impl rlp::Decodable for Multiproof {
-    fn decode(rlp: &rlp::Rlp) -> Result<Self, rlp::DecoderError> {
-        let hashes: Vec<Vec<u8>> = rlp.list_at(0usize)?;
-        let keyvals: Vec<Vec<u8>> = rlp.list_at(1usize)?;
-        let instructions: Vec<Instruction> = rlp.list_at(2usize)?;
-        Ok(Multiproof {
-            hashes: hashes,
-            instructions: instructions,
-            keyvals: keyvals,
-        })
-    }
-}
-
-impl std::fmt::Display for Multiproof {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "instructions:\n")?;
-        for i in self.instructions.iter() {
-            write!(f, "\t{:?}\n", i)?;
-        }
-        write!(f, "keyvals:\n")?;
-        for kv in self.keyvals.iter() {
-            write!(f, "\t{:?}\n", hex::encode(kv))?;
-        }
-        write!(f, "hashes:\n")?;
-        for h in self.hashes.iter() {
-            write!(f, "\t{}\n", hex::encode(h))?;
-        }
-        Ok(())
-    }
-}
 
 // Rebuilds the tree based on the multiproof components
 pub fn rebuild(proof: &Multiproof) -> Result<Node, String> {
@@ -747,6 +385,7 @@ mod tests {
     use super::Instruction::*;
     use super::Node::*;
     use super::*;
+    use sha3::{Digest, Keccak256};
 
     #[test]
     fn validate_tree() {
@@ -1518,118 +1157,6 @@ mod tests {
                     EmptySlot
                 ]))
             )
-        )
-    }
-
-    #[test]
-    fn single_value_hash() {
-        let leaf = Leaf(NibbleKey::from(vec![1, 2, 3]), vec![4, 5, 6]);
-        assert_eq!(leaf.composition(), vec![199, 130, 49, 35, 131, 4, 5, 6]);
-        assert_eq!(
-            leaf.hash(),
-            vec![
-                49, 52, 36, 163, 130, 221, 147, 71, 223, 121, 140, 253, 220, 114, 118, 36, 169, 89,
-                166, 229, 35, 209, 64, 242, 137, 173, 129, 227, 76, 91, 110, 151
-            ]
-        );
-    }
-
-    #[test]
-    fn big_value_single_key_hash() {
-        assert_eq!(
-            Leaf(NibbleKey::from(vec![0u8; 32]), vec![4u8; 32]).hash(),
-            vec![
-                99, 116, 144, 157, 101, 254, 188, 135, 196, 46, 49, 240, 157, 79, 192, 61, 117,
-                243, 84, 131, 36, 12, 147, 251, 17, 134, 48, 59, 76, 39, 205, 106
-            ]
-        );
-    }
-
-    #[test]
-    fn big_value_single_big_key_hash() {
-        assert_eq!(
-            Leaf(NibbleKey::from(vec![0u8; 32]), vec![1u8; 32]).hash(),
-            vec![
-                132, 254, 5, 139, 174, 187, 212, 158, 12, 39, 213, 88, 18, 194, 107, 214, 83, 52,
-                2, 1, 66, 133, 239, 172, 206, 141, 135, 220, 34, 196, 98, 222
-            ]
-        );
-    }
-
-    #[test]
-    fn empty_value_hash() {
-        let node = EmptySlot;
-        assert_eq!(node.composition(), vec![0u8; 0]);
-        assert_eq!(
-            node.hash(),
-            vec![
-                197, 210, 70, 1, 134, 247, 35, 60, 146, 126, 125, 178, 220, 199, 3, 192, 229, 0,
-                182, 83, 202, 130, 39, 59, 123, 250, 216, 4, 93, 133, 164, 112
-            ]
-        );
-    }
-
-    #[test]
-    fn branch_hash() {
-        let branch = Branch(vec![
-            Leaf(NibbleKey::from(vec![]), vec![4, 5, 6]),
-            EmptySlot,
-            Leaf(NibbleKey::from(vec![9]), vec![10, 11, 12]),
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-            EmptySlot,
-        ]);
-        assert_eq!(
-            branch.composition(),
-            vec![
-                219, 197, 32, 131, 4, 5, 6, 128, 197, 57, 131, 10, 11, 12, 128, 128, 128, 128, 128,
-                128, 128, 128, 128, 128, 128, 128, 128, 128
-            ]
-        );
-        assert_eq!(
-            branch.hash(),
-            vec![
-                6, 134, 255, 246, 145, 43, 211, 204, 240, 23, 77, 89, 244, 40, 13, 2, 201, 73, 218,
-                51, 53, 12, 149, 35, 120, 93, 254, 247, 104, 88, 103, 177
-            ]
-        );
-    }
-
-    #[test]
-    fn encode_decode_instruction() {
-        let instructions = vec![LEAF(1), ADD(5), EXTENSION(vec![3u8; 4]), BRANCH(6)];
-
-        let encoded = rlp::encode_list(&instructions);
-        let decoded = rlp::decode_list::<Instruction>(&encoded);
-        assert_eq!(decoded, instructions);
-    }
-
-    #[test]
-    fn encode_decode_multiproof() {
-        let mp = Multiproof {
-            hashes: vec![vec![1u8; 32]],
-            instructions: vec![LEAF(0)],
-            keyvals: vec![rlp::encode(&Leaf(NibbleKey::from(vec![1]), vec![2]))],
-        };
-        let rlp = rlp::encode(&mp);
-        let decoded = rlp::decode::<Multiproof>(&rlp).unwrap();
-        assert_eq!(
-            decoded,
-            Multiproof {
-                hashes: vec![vec![1u8; 32]],
-                keyvals: vec![rlp::encode(&Leaf(NibbleKey::from(vec![1]), vec![2]))],
-                instructions: vec![LEAF(0)]
-            }
         )
     }
